requirements-pdf.txt
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=10.0.0
